apache::vhost-ssl - allow management of ssl cert/key content
Previously $cert, $certkey, $cacert and $certchain were only
references to the name of the cert/key file. The management
of the content of these files was left to the puppeter,
using file resources.

Now these 4 parameters are used to directly define the source
(on puppet's fileserver for instance) of these files.

By default, the files generated by generate-ssl-cert.sh will
be used.