apache: simplified listen/namevirtualhost directives management.
I basically replaced the $apache_ports/$apache_ssl_ports mechanism by
calls to the Apache::Listen and Apache::Namevhost definitions.

If the default behaviour (listen on port 80 for http / port 443 for https)
must be changed, the idea would now be to subclass
apache::base / apache::base::ssl and change the values of the
Apache::Listen and Apache::Namevhost definitions.