apache::vhost $config_file now accepts full source parameter
The previous behaviour was to fetch the vhost config from the files included in
the apache module, which doesn't make any sense since it is supposed to be
completely generic.

This patch changes the behaviour to accept values such as
"puppet:///module/vhost.conf" or "file:///etc/httpd/vhost.conf".